requirements.txt
factory_boy
flask-resty
pytest
python-dotenv
sqlalchemy-utils
httpie
